pyproject.toml
[project]
name = "listing-agent-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Mizael Paredes"}
]
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic (>=2.11.9,<2.13)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "greenlet (>=3.4.0,<4.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "crewai (<=1.12.0)",
    "groq (>=1.2.0,<2.0.0)",
    "litellm (>=1.83.14,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
